[Refactor] 함수명이 중복되어 ~Mode 컴포넌트들을 ~ModeWrap으로 변경 #443
