+++ components/mode/modeWraps/ProfileModeWrap.tsx
@@ -1,24 +1,17 @@
-import React, { useState } from 'react';
+import React from 'react';
+import { useState } from 'react';
 import { useRecoilValue } from 'recoil';
-import { SeasonMode } from 'types/mainType';
 import { seasonListState } from 'utils/recoil/seasons';
-import UserGameSearchBar from '../modeItems/UserGameSearchBar';
 import SeasonDropDown from 'components/mode/modeItems/SeasonDropDown';
-import ModeRadiobox from 'components/mode/modeItems/ModeRadiobox';
 import styles from 'styles/mode/ModeWrap.module.scss';
 
-interface GameModeProps {
+interface ProfileModeWrapProps {
   children: React.ReactNode;
 }
 
-export default function GameMode({ children }: GameModeProps) {
+export default function ProfileModeWrap({ children }: ProfileModeWrapProps) {
   const { seasonList } = useRecoilValue(seasonListState);
   const [season, setSeason] = useState<number>(seasonList[0]?.id);
-  const [radioMode, setRadioMode] = useState<SeasonMode>('both');
-
-  const modeChangeHandler = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setRadioMode(e.target.value as SeasonMode);
-  };
 
   const seasonDropDownHandler = (e: React.ChangeEvent<HTMLSelectElement>) => {
     setSeason(parseInt(e.target.value));
@@ -26,9 +19,8 @@ export default function GameMode({ children }: GameModeProps) {
 
   return (
     <div>
-      <div className={styles.gameModeWrap}>
-        <UserGameSearchBar />
-        {radioMode === 'rank' && seasonList && (
+      <div className={styles.profileModeWrap}>
+        {seasonList && (
           <SeasonDropDown
             seasonList={seasonList}
             value={season}
@@ -36,9 +28,7 @@ export default function GameMode({ children }: GameModeProps) {
           />
         )}
       </div>
-      <ModeRadiobox mode={radioMode} onChange={modeChangeHandler} />
       {React.cloneElement(children as React.ReactElement, {
-        mode: radioMode,
         season,
       })}
     </div>
