+++ pages/match.tsx
@@ -1,14 +1,14 @@
 import MatchBoard from 'components/match/MatchBoard';
-import MatchMode from 'components/mode/modeWraps/MatchMode';
+import MatchModeWrap from 'components/mode/modeWraps/MatchModeWrap';
 import styles from 'styles/match/match.module.scss';
 
 export default function Match() {
   return (
     <div className={styles.container}>
       <h1 className={styles.title}>Match</h1>
-      <MatchMode>
+      <MatchModeWrap>
         <MatchBoard type='single' />
-      </MatchMode>
+      </MatchModeWrap>
     </div>
   );
 }
